requirements.txt
Flask==2.1.0
tensorflow==2.17.0
tensorflow-hub==0.12.0
Pillow==9.2.0
